# bpo-23689: re module, fix memory leak when a match is terminated by a signal or memory allocation failure