--- a/Lib/test/test_re.py
+++ b/Lib/test/test_re.py
@@ -1643,9 +1643,9 @@ class ReTests(unittest.TestCase):
         long_overflow = 2**128
         self.assertRaises(TypeError, re.finditer, "a", {})
         with self.assertRaises(OverflowError):
-            _sre.compile("abc", 0, [long_overflow], 0, {}, ())
+            _sre.compile("abc", 0, [long_overflow], 0, {}, (), 0)
         with self.assertRaises(TypeError):
-            _sre.compile({}, 0, [], 0, [], [])
+            _sre.compile({}, 0, [], 0, [], [], 0)
 
     def test_search_dot_unicode(self):
         self.assertTrue(re.search("123.*-", '123abc-'))