--- a/Misc/NEWS.d/next/Library/2022-03-31-21-08-41.bpo-23689.TFSc3E.rst
+++ b/Misc/NEWS.d/next/Library/2022-03-31-21-08-41.bpo-23689.TFSc3E.rst
@@ -1,0 +1,2 @@
+:mod:`re` module: fix memory leak when a match is terminated by a signal or
+memory allocation failure. Patch by Ma Lin.