@serhiy-storchaka , please review the commits one by one.

MSVC 2022 non-PGO release build, run `test_re.py`. It's a bit slower.
```
(best of 10)
Before: 0.278 sec
After:  0.283 sec
```

But when switching REPEAT very frequently, it has some speed up:
```python
import pyperf

setup = ("import re;"
         "s = 100_000_000 * 'a';"
         "p = r'.*?(?:bb)+';"
         "p = re.compile(p)")
stmt = "p.match(s)"

runner = pyperf.Runner()
runner.timeit("match", stmt=stmt, setup=setup)

# Before: Mean +- std dev: 1.88 sec +- 0.03 sec
# After:  Mean +- std dev: 1.63 sec +- 0.03 sec
```
<!--
Thanks for your contribution!
Please read this comment in its entirety. It's quite important.

# Pull Request title

It should be in the following format:

```
bpo-NNNN: Summary of the changes made
```

Where: bpo-NNNN refers to the issue number in the https://bugs.python.org.

Most PRs will require an issue number. Trivial changes, like fixing a typo, do not need an issue.

# Backport Pull Request title

If this is a backport PR (PR made against branches other than `main`),
please ensure that the PR title is in the following format:

```
[X.Y] <title from the original PR> (GH-NNNN)
```

Where: [X.Y] is the branch name, e.g. [3.6].

GH-NNNN refers to the PR number from `main`.

-->


<!-- issue-number: [bpo-23689](https://bugs.python.org/issue23689) -->
https://bugs.python.org/issue23689
<!-- /issue-number -->